Infected A중간점검 이동공격시 사거리 조절이 약간안되는데 잠자기 전까지 잡아놓으려고함

+++ Reference/Imgui/imNeoSequencer/imgui_neo_sequencer.cpp
@@ -1041,7 +1041,7 @@ namespace ImGui
         for (uint32_t i = 0; i < keyframes.size(); i++)
             c_keyframes[i] = &keyframes[i];
 
-        return BeginNeoTimeline(label, c_keyframes.data(), c_keyframes.size(), open, flags);
+        return BeginNeoTimeline(label, c_keyframes.data(), (uint32_t)c_keyframes.size(), open, flags);
     }
 
 #endif
